Shared scheme added + unwanted files removed

+++ CPAudioPlayer/CPBandEqulizer.m
@@ -24,7 +24,7 @@
     self = [super init];
     if (self) {
         bandEQunit = bandUnit;
-        self.numBands = frequency.count;
+        self.numBands = (UInt32)frequency.count;
         self.bands = frequency;
     }
     return self;
@@ -75,7 +75,7 @@
 -(void)setBands:(NSArray *)bands
 {
     _bands = bands;
-    [self setMaxNumberOfBands:_bands.count];
+    [self setMaxNumberOfBands:(UInt32)_bands.count];
     for (UInt32 i=0; i<bands.count; i++) {
         AudioUnitSetParameter(bandEQunit,
                               kAUNBandEQParam_Frequency+i,
@@ -105,7 +105,7 @@
 - (AudioUnitParameterValue)gainForBandAtPosition:(NSUInteger)bandPosition
 {
     AudioUnitParameterValue gain;
-    AudioUnitParameterID parameterID = kAUNBandEQParam_Gain + bandPosition;
+    AudioUnitParameterID parameterID = kAUNBandEQParam_Gain + (UInt32)bandPosition;
     
     AudioUnitGetParameter(bandEQunit,
                           parameterID,
@@ -117,7 +117,7 @@
 
 -(void)setGainForBandAtPosition:(NSInteger)bandPosition value:(float)gain
 {
-    AudioUnitParameterID parameterID = (UInt32) kAUNBandEQParam_Gain + bandPosition;
+    AudioUnitParameterID parameterID = (UInt32) kAUNBandEQParam_Gain + (UInt32)bandPosition;
     AudioUnitSetParameter(bandEQunit,
                           parameterID,
                           kAudioUnitScope_Global,
